merge main into tony

+++ DreamGuard.BE.BLL/Requests/VnPaymentRequest.cs
@@ -10,7 +10,7 @@ namespace DreamGuard.BE.BLL.Requests
         public required string PaymentId { get; set; }
         public required string OrderCode { get; set; }
         public required string Description { get; set; }
-        public required int Amount { get; set; }
+        public required decimal Amount { get; set; }
         public required string IpAddress { get; set; }
         public DateTime CreatedDate { get; set; }
     }
